+++ packages/two_dimensional_scrollables/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fixes hit testing for `TreeView` row content and gestures after horizontal scrolling.
 * Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
+* Fixes an issue where pinned cells were losing hit tests to underlying non-pinned cells.
 
 ## 0.5.2